\encoding{UTF-8}
\name{aoscp}
\alias{aoscp}
\alias{aoscp1}
\title{Plots from oxidation state of carbon paper}
\description{
Plots from the paper by Dick (2014).
}

\usage{
  aoscp1(pdf = FALSE)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{aoscp1} \tab \ZC of amino acids vs \ZC of RNA codons and hydropathy index of amino acids \cr
}
}

\references{
Dick, J. M. (2014) Average oxidation state of carbon in proteins. \emph{J. R. Soc. Interface} \bold{11}, 20131095. \doi{10.1098/rsif.2013.1095}
}

\examples{
aoscp1()
}
