\encoding{UTF-8}
\name{cpcp}
\alias{cpcp}
\title{Plots from cancer potential diagrams papers}
\description{
Plots from the papers by Dick (2016 and 2017).
}

\details{
There are no functions for these papers.
}

\section{Files in extdata/Dic16}{
  \describe{
    \item{\code{microbes.csv}, \code{microbial.aa.csv}}{
      Data for microbial occurrence in colorectal cancer and normal tissue, and overall amino acid compositions of proteins in microbial species.
      The files are from the Supporting Information of Dick (2016).
    }
  }
}

\section{Files in extdata/cpcp}{
  \describe{
    \item{\code{colorectal.Rmd}, \code{hyperosmotic.Rmd}, \code{hypoxia.rmd}, \code{pancreatic.Rmd}}{
      Vignettes showing compositional analysis of differential protein expression, based on Figures 1-2 and Tables 1-4 of Dick (2017).
      Updated versions of these plots and the underlying datasets are in Dick (2020).
    }
  }
}

\references{
Dick, J. M. (2016) Proteomic indicators of oxidation and hydration state in colorectal cancer. \emph{PeerJ} \bold{4}, e2238. \doi{10.7717/peerj.2238}

Dick, J. M. (2017) Chemical composition and the potential for proteomic transformation in cancer, hypoxia, and hyperosmotic stress. \emph{PeerJ} \bold{5}, e3421 \doi{10.7717/peerj.3421}

Dick, J. M. (2020) Water as a reactant in the differential expression of proteins in cancer. \emph{bioRxiv}. \doi{10.1101/2020.04.09.035022}
}
