\encoding{UTF-8}
\name{diffexpr}
\alias{diffexpr}
\alias{mkvig}
\alias{check_IDs}
\alias{cleanup}
\alias{get_comptab}
\alias{diffplot}
\alias{qdist}
\alias{xsummary}
\alias{xsummary2}
\alias{xsummary3}
\title{Functions for processing differential expression datasets}
\description{
  These functions were moved from the canprot package in February 2024.
}

\usage{
  mkvig(vig = NULL)
  check_IDs(dat, IDcol, aa_file = NULL, updates_file = NULL)
  cleanup(dat, IDcol, up2 = NULL)
  get_comptab(pdat, var1 = "Zc", var2 = "nH2O", plot.it = FALSE,
    mfun = "median", oldstyle = FALSE)
  diffplot(comptab, vars = c("Zc", "nH2O"), col = "black", plot.rect = FALSE,
           pt.text = c(letters, LETTERS), cex.text = 0.85, oldstyle = FALSE,
           pch = 1, cex = 2.1, contour = TRUE, col.contour = par("fg"),
           probs = 0.5, add = FALSE, labtext = NULL, ...)
  qdist(pdat, vars = c("Zc", "nH2O"), show.steps = FALSE)
  xsummary(comptab, vars = c("Zc", "nH2O"))
  xsummary2(comptab1, comptab2)
  xsummary3(comptab1, comptab2, comptab3)
}

\arguments{
  \item{vig}{character, name of a vignette without \samp{.Rmd} extension}
  \item{dat}{data frame, protein expression data}
  \item{IDcol}{character, name of column that has the UniProt IDs}
  \item{aa_file}{character, name of file with additional amino acid compositions}
  \item{updates_file}{character, name of file with old to new ID mappings}
  \item{up2}{logical, TRUE for up-regulated proteins, FALSE for down-regulated proteins}
  \item{pdat}{list, data object generated by a \code{\link{pdat_}} function}
  \item{var1}{character, the first variable}
  \item{var2}{character, the second variable}
  \item{plot.it}{logical, make a scatterplot?}
  \item{mfun}{character, either \samp{median} or \samp{mean}}
  \item{oldstyle}{logical, also calculate \code{\link[canprot]{CLES}} and \emph{p}-values?}
  \item{comptab}{list or data frame, output of \code{get_comptab}}
  \item{vars}{character, which variables (chemical metrics) to calculate or plot}
  \item{col}{character or numeric, color(s) for the points}
  \item{plot.rect}{logical, plot a reference rectangle?}
  \item{pt.text}{character, text labels for the points}
  \item{cex.text}{numeric, size of text labels}
  \item{pch}{numeric, point symbol}
  \item{cex}{numeric, point size}
  \item{contour}{logical, add contour lines?}
  \item{col.contour}{character or numeric, color of contour lines}
  \item{probs}{numeric, probability level(s) for contours}
  \item{add}{logical, add to an existing plot?}
  \item{labtext}{character, text to add to axis labels}
  \item{...}{other argumenents passed to \code{\link{plot}}}
  \item{show.steps}{logical, show the steps using \code{\link{plot.ecdf}}?}
  \item{comptab1}{list, output of \code{get_comptab}}
  \item{comptab2}{list, output of \code{get_comptab}}
  \item{comptab3}{list, output of \code{get_comptab}}
}

\details{

These functions for processing differential expression datasets for the gradH2O and canH2O papers were previously in the canprot package.

\code{mkvig} compiles the indicated vignette for chemical analysis of differential expression datasets and opens it in the browser.
Pandoc (including pandoc-citeproc), as a system dependency of \pkg{rmarkdown}, must be installed.
See \pkg{rmarkdown}'s \samp{pandoc} vignette for installation tips.
The vignettes can also be run using e.g. \code{\link{demo}("HPA")}, and through the interactive help system (\code{\link{help.start}} > Packages > JMDplots > Code demos).
The available vignettes are listed here:
\itemize{
    \item \emph{Cell culture} -- \samp{hypoxia}, \samp{secreted}, \samp{osmotic_bact}, \samp{osmotic_euk}, \samp{osmotic_halo}, \samp{glucose}, \samp{3D}, \samp{osmotic_gene}, \samp{yeast_stress}
    \item \emph{Cancer} -- \samp{breast}, \samp{colorectal}, \samp{liver}, \samp{lung}, \samp{pancreatic}, \samp{prostate}
    \item \emph{Pan-cancer} -- \samp{TCGA}, \samp{HPA}
}

\code{check_IDs} is used to check for known UniProt IDs and to update obsolete IDs.
The source IDs should be provided in the \code{IDcol} column of \code{dat}; multiple IDs for one protein can be separated by a semicolon.
The function keeps the first \dQuote{known} ID for each protein, which must be present in one of these groups:
\itemize{
  \item The \code{\link[canprot]{human.aa}} dataset of amino acid compositions.
  \item Old UniProt IDs that are mapped to new UniProt IDs in \file{extdata/diffexpr/uniprot_updates.csv} or in \code{updates_file} if specified.
  \item IDs of proteins in \code{aa_file}, which lists amino acid compositions in the format described for \code{\link[canprot]{human.aa}} (see \code{\link[CHNOSZ]{thermo}$protein} for details).
}

\code{cleanup} removes proteins with unavailable IDs, ambiguous expression ratios, and duplicated IDs.
\code{IDcol} is the name of the column that has the UniProt IDs, and \code{up2} indicates the expression change for each protein.
The function removes proteins with unavailable (NA or "") or duplicated IDs.
If \code{up2} is provided, the function also removes unquantified proteins (those that have NA values of \code{up2}) and those with ambiguous expression ratios (up and down for the same ID).
For each operation, a message is printed describing the number of proteins that are \samp{unavailable}, \samp{unquantified}, \samp{ambiguous}, or \samp{duplicated}.
Alternatively, if \code{IDcol} is a logical value, it selects proteins to be unconditionally removed.

\code{get_comptab} computes differences of chemical metrics between groups of up- and down-regulated proteins.
\itemize{
\item{Differentially expressed proteins are identified by the value of \code{pdat$up2} (TRUE for up-regulated proteins and FALSE for down-regulated proteins).}
\item{The differences are calculated as (median for up-regulated proteins) - (median for down-regulated proteins).
If \code{mfun} is \samp{mean}, means of the groups are used instead.}
\item{If \code{oldstyle} is TRUE, the function also calculates the common language effect size (\code{\link[canprot]{CLES}}, in percent) and \emph{p}-value for each variable.}
\item{Volume is calculated using amino acid group additivity as described by Dick et al. (2006).}
\item{Set \code{plot.it} to \code{TRUE} to make a scatterplot.
Open red squares and filled blue circles stand for up-regulated and down-regulated proteins, respectively.}
}

\code{diffplot} makes a plot with points showing the differences between up- and down-regulated proteins for two chemical metrics, as calculated by \code{\link{get_comptab}}.
\itemize{
\item{The default setting of \code{vars} refers to average oxidation state of carbon (\Zc) as the x-variable and stoichiometric hydration state (\nH2O) as the y-variable.}
\item{The colors of the points are controlled by \code{col}, which is recycled to be equal to the number of comparisons in \code{comptab}.}
\item{If \code{plot.rect} is TRUE, a shaded \code{\link{rect}}angle is drawn with coordinates -0.01, -0.01, 0.01, 0.01.
This is useful for visualizing the different scales of multi-panel plots.}
\item{If \code{pt.text} is not NA or FALSE, \code{\link{text}} labels are added with size controlled by \code{cex.text}.
The default value produces labels that are taken sequentially from the 26 lowercase Roman letters in alphabetical order (\code{\link{letters}}), followed by the set of uppercase letters (\code{\link{LETTERS}}).}
\item{For \code{labtext} = NULL, descriptive text (\dQuote{median difference} or \dQuote{mean difference}) is added to the axis labels in parentheses.
This text can be changed by giving a value in \code{labtext} (for both axes), two values (for each axis), or NA to suppress the text.}
\item{\code{cplab} is a list of formatted labels used by \code{diffplot}.
It is an exported object, available to the user and other packages.}
}

\code{qdist} makes a quantile distribution plot with lines for both up- and down-regulated proteins.
The variable (\code{var}) can be \samp{Zc}, \samp{H2O}, or both (two plots are made for the latter).
The horizontal axis is the variable and the vertical axis is the quantile point.
A solid black line is drawn for the down-regulated proteins, and a dashed red line for the up-regulated proteins.
The median difference is shown by a gray horizontal line drawn between the distributions at the 0.5 quantile point.

\code{xsummary} makes an HTML table summarizing chemical differences using \code{\link[xtable]{xtable}}.
Bold and underline formatting is used to highlight significant chemical differences.
The \emph{p}-value is bolded if it is less than 0.05, and the percent common language effect size (\code{\link[canprot]{CLES}}) is bolded if it is <= 40 or >= 60.
The mean (or median) difference is [underlined / bolded] if [only one of / both] the \emph{p}-value and CLES pass these cutoffs.
The generated table is written to the console, and can be used in a vignette using the \code{results = "asis"} chunk option.

\code{xsummary2} is an updated version that is used in the current vignettes in the package.
It shows negative numbers in bold (\emph{p}-value and CLES are not shown).
\code{xsummary3} is a further revision that shows GRAVY and pI; it is used in the \samp{osmotic_bact} and \samp{osmotic_halo} vignettes.

}

\section{Plot style}{
The overall style of the plot is controlled by \code{oldstyle}.
\describe{
  \item{\code{oldstyle = FALSE}}{
    This is the current default style.
    Use \code{pch} and \code{cex} to control the point symbol and size.
    Contours are added for confidence regions of highest probability density, computed using a 2-D kernel density estimate (\code{\link[MASS]{kde2d}}).
    \code{probs} gives the probability level(s) and \code{col.contour} sets the color(s) of the contour lines.
    \code{contour} can be a logical vector, indicating which points to include; set it to FALSE to omit the contour lines.

    The code to calculate the contour levels is modified from \code{HPDregionplot} in the \pkg{emdbook} package by Ben Bolker (\url{https://cran.r-project.org/package=emdbook}).
  }
  \item{\code{oldstyle = TRUE}}{
    This style was used for the historical (2017) vignettes, which have been moved to the \samp{extdata/cpcp} directory in \pkg{JMDplots} (\url{https://github.com/jedick/JMDplots}).
    For each dataset, the point symbol is a filled square if the \emph{p}-values of both the x-variable and y-variable are less than 0.05, a filled circle if the \emph{p}-value of one of the x- or y-variables is less than 0.05, and an open circle otherwise.
    A solid line is drawn from the point to the corresponding axis if the rounded, absolute value of (\code{\link[canprot]{CLES}} in percent - 50) of the x- or y-variable is greater than or equal 10.
    Otherwise, a dashed line is drawn from the point to the corresponding axis if the \emph{p}-value of the x- or y-variable is less than 0.05.
    Otherwise, no line is drawn.
  }
}
}

\value{
For \code{check_IDs}, \code{dat} is returned with possibly changed values in the column designated by \code{IDcol}; old IDs are replaced with new ones, the first known ID for each protein is kept, then proteins with no known IDs are assigned \code{NA}.

For \code{get_comptab}, a data frame is returned invisibly containing the columns \samp{dataset}, \samp{description}, \samp{n1} (number of down-regulated proteins), \samp{n2} (number of up-regulated proteins), followed two sets of columns for the variables.
These are denoted generically as (\samp{var.mfun1}, \samp{var.mfun2}, \samp{var.diff}, \samp{var.CLES}, \samp{var.p.value}), where \samp{var} is replaced by the name of \code{var1} or \code{var2}, and \samp{mfun} is replaced by the value of \code{mfun}.
For example, \samp{Zc.median1} and \samp{Zc.median2} are the median \Zc of the down- and up-regulated proteins, respectively.

For \code{xsummary}, (invisibly) the data frame used to make the table; this data frame differs from \code{comptab} by having row names added (alphabetical one-letter IDs for the datasets).
}


\seealso{
\code{\link{pdat_}}
}


\examples{
\dontrun{
mkvig("osmotic_gene")
}

# Synthetic data to show actions for incorrect IDs
ID <- c("P61247;PXXXXX", "PYYYYY;P46777;P60174", "PZZZZZ")
dat <- data.frame(ID = ID, stringsAsFactors = FALSE)
# Get the first known ID for each protein; the third one is NA
check_IDs(dat, "ID")
# Update an old ID
dat <- data.frame(Entry = "P50224", stringsAsFactors = FALSE)
check_IDs(dat, "Entry")

# Set up a workflow to clean data and retrieve amino acid compositions
extdatadir <- system.file("extdata", package = "JMDplots")
datadir <- paste0(extdatadir, "/diffexpr/pancreatic/")
dataset <- "CYD+05"
dat <- read.csv(paste0(datadir, dataset, ".csv.xz"), as.is = TRUE)
up2 <- dat$Ratio..cancer.normal. > 1
# Remove two unavailable and one duplicated proteins
dat <- cleanup(dat, "Entry", up2)
# Now we can retrieve the amino acid compositions
aa <- canprot::human_aa(dat$Entry)

# Read another data file
datadir <- paste0(system.file("extdata", package = "JMDplots"), "/diffexpr/colorectal/")
dataset <- "STK+15"
dat <- read.csv(paste0(datadir, "STK+15.csv.xz"), as.is = TRUE)
# Remove unavailable proteins
dat <- cleanup(dat, "uniprot")
# Remove proteins that have less than 2-fold expression ratio
dat <- cleanup(dat, abs(log2(dat$invratio)) < 1)

# Analysis of differential expression data
pd <- pdat_colorectal("JKMF10")
# Default variables: Zc and nH2O
get_comptab(pd, plot.it = TRUE)
# Protein length and per-residue volume
get_comptab(pd, "nAA", "V0", plot.it = TRUE)

# Make an old-style plot for two datasets
comptab <- lapply(c("JKMF10", "WDO+15_C.N"), function(dataset) {
  pdat <- pdat_colorectal(dataset)
  get_comptab(pdat, oldstyle = TRUE)
})
diffplot(comptab, oldstyle = TRUE)

# Plot the data of Jimenez et al., 2010 for colorectal cancer
pdat <- pdat_colorectal("JKMF10")
qdist(pdat)

# Making a table
comptab <- lapply(c("JKMF10", "WDO+15_C.N"), function(dataset) {
  pdat <- pdat_colorectal(dataset)
  get_comptab(pdat, oldstyle = TRUE)
})
xsummary(comptab)

}

\references{
Dick, J. M., LaRowe, D. E. and Helgeson, H. C. (2006) Temperature, pressure, and electrochemical constraints on protein speciation: Group additivity calculation of the standard molal thermodynamic properties of ionized unfolded proteins. \emph{Biogeosciences} \bold{3}, 311--336. \doi{10.5194/bg-3-311-2006}

Jimenez, C. R. and Knol, J. C. and Meijer, G. A. and Fijneman, R. J. A. (2010) Proteomics of colorectal cancer: Overview of discovery studies and identification of commonly identified cancer-associated proteins and candidate CRC serum markers. \emph{J. Proteomics} \bold{73}, 1873--1895. \doi{10.1016/j.jprot.2010.06.004}
}

