\encoding{UTF-8}
\name{scsc}
\alias{scsc}
\alias{scsc2}
\alias{scsc3}
\title{Plots from yeast subcellular location paper}
\description{
Plots from the paper by Dick (2009).
}

\usage{
  scsc2()
  scsc3()
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{scsc2} \tab Oxygen fugacity - activity of H2O predominance diagrams \cr
  \code{scsc3} \tab Relative equilibrium abundances of proteins as a function of oxygen fugacity \cr
}
}

\references{
Dick, J. M. (2009) Calculation of the relative metastabilities of proteins in subcellular compartments of \emph{Saccharomyces cerevisiae}. \emph{BMC Syst. Biol.} \bold{3},75. \doi{10.1186/1752-0509-3-75}
}

\examples{
scsc3()
}
