\encoding{UTF-8}
\name{canH2O}
\alias{canH2O}
\alias{canH2O1}
\alias{canH2O2}
\alias{canH2O3}
\title{Plots from water in cancer paper}
\description{
Plots from the paper by Dick (2020).
}

\usage{
  canH2O1(pdf = FALSE)
  canH2O2(pdf = FALSE)
  canH2O3(pdf = FALSE)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{canH2O1} \tab Study overview \cr
  \code{canH2O2} \tab Median differences of protein length, \nH2O and \ZC for cell culture and cancer \cr
  \code{canH2O3} \tab \nH2O-\ZC and phylostrata-\nAA plots for TCGA and HPA datasets \cr
}

}

\references{
Dick, J. M. (2020) Water as a reactant in the differential expression of proteins in cancer. \emph{(preprint)}
}

\examples{
canH2O1()
}
