\encoding{UTF-8}
\name{canH2O}
\alias{canH2O}
\alias{canH2O1}
\alias{canH2O2}
\alias{canH2O3}
\alias{canH2O4}
\alias{canH2O5}
\alias{canH2OT1}
\alias{canH2OT2}
\alias{canH2OS1}
\alias{canH2OS2}
\title{Plots from water in cancer paper}
\description{
Plots from the paper by Dick (2020).
Function names ending with \samp{T*} and \samp{S*} refer to supplementary tables and figures.
}

\usage{
  canH2O1(pdf = FALSE)
  canH2O2(pdf = FALSE)
  canH2O3(pdf = FALSE)
  canH2O4(pdf = FALSE)
  canH2OT1()
  canH2OT2()
  canH2OS1(pdf = FALSE)
  canH2OS2(pdf = FALSE)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{canH2O1} \tab Study overview \cr
  \code{canH2O2} \tab Median differences of protein length, \nH2O and \ZC for cell culture and cancer tissue \cr
  \code{canH2O3} \tab \nH2O-\ZC plots for TCGA and HPA datasets \cr
  \code{canH2O4} \tab Differentially expressed genes in aneuploid and osmotically shocked yeast cells \cr
  \code{canH2OT1} \tab Stoichiometric matrix for amino acids with QEC basis species \cr
  \code{canH2OT2} \tab Mean differences and \emph{p}-values for all datasets in each condition \cr
  \code{canH2OS1} \tab \nO2-\ZC and \nH2O-\ZC correlations using QEC basis species \cr
  \code{canH2OS2} \tab HPA-TCGA scatterplots for \ZC and \nH2O \cr
}
}

\section{Acknowledgements}{
The results shown here are in part based upon data generated by the TCGA Research Network (\url{https://www.cancer.gov/tcga}) and Human Protein Atlas (\url{https://www.proteinatlas.org/}).
}

\references{
Dick, J. M. (2020) Water as a reactant in the differential expression of proteins in cancer. \emph{Comp. Sys. Onco.} \bold{1}:e1007 \doi{10.1002/cso2.1007}
}

\examples{
canH2O1()
}
