\encoding{UTF-8}
\name{chem16S}
\alias{chem16S}
\alias{getmdat}
\alias{getRDP}
\alias{getmap}
\alias{getmetrics}
\title{Chemical analysis of proteomes inferred from 16S sequences}
\description{
Functions to calculate chemical parameters (\ZC and \nH2O) for inferred microbial community proteomes.
Proteomes are inferred by combining taxonomic classifications of 16S rRNA gene sequences with RefSeq microbial proteomes.
}

\usage{
  getmdat(study, dropNA = TRUE)
  getRDP(study, cn = FALSE, mdat = NULL, lineage = NULL)
  getmap(study, RDP = NULL, lineage = NULL)
  getmetrics(study, cn = FALSE, mdat = NULL, RDP = NULL, map = NULL,
    lineage = NULL, metrics = NULL, groups = NULL)
}

\arguments{
  \item{study}{character, study name}
  \item{dropNA}{logical, exclude samples with NA name in metadata?}
  \item{cn}{logical, adjust counts for 16S rRNA gene copy number?}
  \item{lineage}{character, regular expression for filtering lineages}
  \item{mdat}{data frame, output from \code{getmdat} (used for memoization by other functions)}
  \item{RDP}{data frame, output from \code{getRDP} (used for memoization by other functions)}
  \item{map}{data frame, output from \code{getmap} (used for memoization by other functions)}
  \item{metrics}{data frame, content of \code{extdata/chem16S/RefSeq_metrics.csv} (used for memoization by other functions)}
  \item{groups}{list, indexing vectors to aggregate groups of samples into one or more meta-samples}
}

\details{
\code{getmdat} gets metadata for a study; the \code{study} argument corresponds to file names in the \code{extdata/chem16S/metadata} directory, without the \code{.csv} suffix.
For some studies, subsets of data are indicated by appending an underscore and suffix to the study name.
The function appends columns for plot parameters (\code{pch}, \code{col}).
For some studies, columns for \code{minuend} and \code{subtrahend} are also provided, to be used for identifying sample pairs for difference calculations.
The default for \code{dropNA} means to exclude samples with NA name in the metadata file, which is used to exclude outliers from the analysis.

\code{getRDP} gets RDP results for all samples in a study.
The output has columns for \samp{lineage}, \samp{rank}, \samp{name}, and columns of sequence classification counts for each sample in the study.
Only rows (lineages) with count > 0 for at least one sample are retained.
Sequences with classification results at only the root or domain level (Root, Archaea, or Bacteria) are omitted because they provide poor taxonomic resolution.
Sequences classified to the class Chloroplast or genera Chlorophyta or Bacillariophyta are also omitted because they have little correspondence with the NCBI taxonomy.
All remaining sequences (those classified to genus or higher levels) can be used for mapping to the NCBI taxonomy.

The location of the \code{chem16S} directory used in \code{getmdat} and \code{getRDP} can be changed by setting the option \code{chem16Sdir}.

\code{getmap} maps RDP names to the NCBI taxonomy.
RDP group names are made by combining the rank and name with an underscore separator (e.g. \samp{genus_Escherichia/Shigella}).
Some particular group names are changed to correspond with the NCBI taxonomy as follows:
\tabular{ll}{
    \strong{RDP} \tab \strong{NCBI} \cr
    genus_Escherichia/Shigella \tab genus_Escherichia \cr
    phylum_Cyanobacteria/Chloroplast \tab phylum_Cyanobacteria \cr
    genus_Marinimicrobia_genera_incertae_sedis \tab species_Candidatus Marinimicrobia bacterium \cr
    class_Cyanobacteria \tab phylum_Cyanobacteria \cr
    genus_Spartobacteria_genera_incertae_sedis \tab species_Spartobacteria bacterium LR76 \cr
    class_Planctomycetacia \tab class_Planctomycetia \cr
    class_Actinobacteria \tab phylum_Actinobacteria \cr
    order_Rhizobiales \tab order_Hyphomicrobiales \cr
    genus_Gp1 \tab genus_Acidobacterium \cr
    genus_Gp6 \tab genus_Luteitalea \cr
    genus_GpI \tab genus_Nostoc \cr
    genus_GpIIa \tab genus_Synechococcus \cr
    genus_GpVI \tab genus_Pseudanabaena \cr
    family_Family II \tab family_Synechococcaceae \cr
    genus_Subdivision3_genera_incertae_sedis \tab family_Verrucomicrobia subdivision 3
}
The group names are then matched to the NCBI taxa listed in \code{extdata/chem16S/groupAA.csv}.
The \code{protein} and \code{organism} columns in this file hold the rank and taxon name extracted from the RefSeq database.
Only exactly matching names are mapped; any unmatching names are excluded from the mapping, and the most abundant unmapped groups are printed.
The return value consists of rownumbers in the data frame generated by reading \code{groupAA.csv}.

\code{getmetrics} calculates the carbon oxidation state (\ZC) and stoichiometric hydration state (\nH2O) for the inferred microbial proteome in each sample.
The inferred proteome can be conceived as the average of amino acid compositions of RefSeq proteomes, weighted by the RDP classification counts; however, the use of precomputed \ZC and \nH2O for each RefSeq proteome (stored in \code{RefSeq_metrics.csv}) is used to speed up the calculation.
For \nH2O, the RDP counts are multiplied by the precomputed \nH2O for the mapped RefSeq group, and the sum is divided by the total number of RDP counts.
For \ZC, the means are further weighted by number of carbon atoms (Dick et al., 2020).

}

\section{Files in extdata/chem16S}{
  \describe{
    \item{\code{pipeline.R}}{Functions to run external programs for sequence analysis (fastq-dump, vsearch, seqtk, RDP Classifier).}
    \item{\code{metadata/*.csv}}{Sample metadata for each study.}
    \item{\code{process.R}}{Functions to create the processed data files listed below:}
    \item{\code{RDP/*.csv.xz}}{RDP Classifier results combined into a single CSV file for each study. (\code{process.R:mkRDP})}
    \item{\code{groupAA.csv}}{Average amino acid composition of RefSeq proteomes at genus and higher ranks. (\code{process.R:mkAA})}
    \item{\code{RefSeq_metrics.csv}}{Chemical parameters (\ZC and \nH2O) for each RefSeq group. (\code{process.R:mkmetrics})}
    \item{\code{bergeyTrainingTree.csv}}{Data from \code{bergeyTrainingTree.xml} in the RDP Classifier package version 2.1.3 (\url{https://sourceforge.net/projects/rdp-classifier/}), used for copy-number adjustment.}
  }
}

\seealso{
\code{\link{chem16S_plot}} for functions used to plot the chemical parameters.
}

\references{
Dick, J. M., Yu, M. and Tan, J. (2020) Uncovering chemical signatures of salinity gradients through compositional analysis of protein sequences. \emph{Biogeosciences} \bold{17}, 6145--6162. \doi{10.5194/bg-17-6145-2020}
}

\examples{
# List available studies
mdatdir <- system.file("extdata/chem16S/metadata", package = "JMDplots")
gsub(".csv", "", dir(mdatdir))

# Get metadata, RDP counts, RDP-NCBI mapping,
# and chemical parameters for one study
getmdat("BGPF13")
getRDP("BGPF13")
getmap("BGPF13")
getmetrics("BGPF13")

# Calculate parameters for aggregated samples for Archaea and Bacteria
mdat <- getmdat("BGPF13")
groups <- list(mdat$cohort == "Archaea", mdat$cohort == "Bacteria")
getmetrics("BGPF13", groups = groups)
}
