\encoding{UTF-8}
\name{evdevH2O}
\alias{evdevH2O}
\alias{evdevH2O1}
\alias{evdevH2O2}
\title{Plots from water loss in evo-devo paper}
\description{
Plots from the paper by Dick (2021).
}

\usage{
  evdevH2O1(pdf = FALSE)
  evdevH2O2(pdf = FALSE)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{evdevH2O1} \tab Compositional analysis of Trigos and Liebeskind datasets \cr
  \code{evdevH2O2} \tab Strategy for finding optimal \logaH2O and \logfO2 for model proteins \cr
}
}

\references{
Dick, J. M. (2021) Redox potential linked to water loss from proteins in evolution and developent. (Preprint)
}

\examples{
evdevH2O1()
}
