\encoding{UTF-8}
\name{genoGOE}
\alias{genoGOE}
\alias{genoGOE_1}
\title{Genomic record of Earth's surface oxygenation}
\description{
Plots from the manuscript by Dick (2024).
}

\usage{
  genoGOE_1(pdf = FALSE, metric = "Zc")
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
  \item{metric}{character, which metric to plot (see \code{\link[canprot]{metrics}})}
}

\details{

\tabular{ll}{
  \code{genoGOE_1} \tab Carbon oxidation state of proteins as a function of gene age in two lineages \cr
}
}

\references{
Dick J (2024) Manuscript in preparation.
}

\examples{
genoGOE_1()
}
