\encoding{UTF-8}
\name{geo16S}
\alias{geo16S}
\alias{geo16S1}
\alias{geo16S2}
\alias{geo16S3}
\alias{geo16S4}
\title{Plots from geobiochemistry of 16S sequences paper}
\description{
Plots from the paper by Dick and Tan (2021).
}

\usage{
  geo16S1(pdf = FALSE)
  geo16S2(pdf = FALSE)
  geo16S3(pdf = FALSE)
  geo16S4(pdf = FALSE)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
}

\details{
This table gives a brief description of each plotting function.

\tabular{ll}{
  \code{geo16S1} \tab Chemical compositions of taxonomic groups \cr
  \code{geo16S2} \tab Natural environment datasets \cr
  \code{geo16S3} \tab Shale gas datasets \cr
  \code{geo16S4} \tab Taxonomic levels \cr
}

}

\examples{
geo16S1()
}
