\encoding{UTF-8}
\name{gradH2O}
\alias{gradH2O}
\alias{gradH2O1}
\title{Plots from salinity gradients paper}
\description{
Plots from the paper by Dick et al. (2019).
}

\usage{
  gradH2O1(pdf = FALSE)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{gradH2O1} \tab Some general characteristics of \ZC of DNA, RNA, and proteins \cr
}
}

\references{
Dick, J. M., Tan, J. and Yu, M. (2019) Geochemistry of metagenomes: Distinct compositional signatures of environmental redox and salinity gradients. \emph{(preprint)}
}

\examples{
gradH2O1()
}
