\encoding{UTF-8}
\name{gradox}
\alias{gradox}
\alias{gradox1}
\alias{gradox2}
\alias{gradoxS1}
\alias{gradoxS2}
\alias{gradox3}
\alias{gradox4}
\alias{gradox5}
\title{Plots from redox gradients paper}
\description{
Plots from the paper by Dick et al. (2019).
}

\usage{
  gradox1(pdf = FALSE)
  gradox2(pdf = FALSE)
  gradoxS1(pdf = FALSE)
  gradoxS2(pdf = FALSE)
  gradox3(mout, pout, pdf = FALSE)
  gradox4(mout, pout, pdf = FALSE)
  gradox5(pdf = FALSE, maxdepth = 500)
}

\arguments{
  \item{pdf}{logical, make a PDF file?}
  \item{mout}{list, value returned by \code{gradoxS1} (DNA \ZC)}
  \item{pout}{list, value returned by \code{gradoxS2} (protein \ZC)}
  \item{maxdepth}{numeric, maximum sample depth (meters)}
}

\details{
This table gives a brief description of each function.

\tabular{ll}{
  \code{gradox1} \tab Some general characteristics of \ZC of DNA, RNA, and proteins \cr
  \code{gradox2} \tab Selected plots of DNA, RNA, and protein \ZC \cr
  \code{gradoxS1} \tab All plots of DNA and RNA \ZC \cr
  \code{gradoxS2} \tab All plots of protein \ZC \cr
  \code{gradox3} \tab \ZC of proteins vs DNA (metagenomes and metatranscriptomes) \cr
  \code{gradox4} \tab Thermodynamic calculations of relative stabilities along redox gradients \cr
  \code{gradox5} \tab \ZC of reads classified to selected abundant species \cr
}
}

\section{Files in extdata/gradox}{
  The files were generated by processing source FASTA files using the "ARAST" workflow; see the paper (Dick et al., 2019) and the Zenodo record (\doi{10.5281/zenodo.2314933}) for more information.
  In the first three directories, files whose names contain "MGD", "MGR", or "MGP" are derived from metegenomic datasets; those with "MTD", "MTR", or "MTP" are derived from metatranscriptomic datasets.

  \describe{
    \item{\code{MGD}}{DNA nucleobase compositions for groups of sequences sampled from source FASTA files that have been trimmed, normalized, and dereplicated. For metatranscriptomes, sequences classified as rRNA by FragGeneScan (Rho et al., 2010) were removed.}
    \item{\code{MGR}}{RNA nucleobase compositions for groups of sequences sampled from coding DNA FASTA files generated by FragGeneScan.}
    \item{\code{MGP}}{Protein amino acid compositions for groups of sequences sampled from protein FASTA files generated by FragGeneScan.}
    \item{\code{one_percent}}{Files generated by using Kraken (Wood and Salzberg, 2014) report files to calculate taxon abundances as a percentage of the number of classified reads, then extracting species that make up at least 1 percent of the classified reads.}
    \item{\code{AA_codon.csv}}{Codon usage data for \emph{Prochlorococcus marinus} str. AS9601 and \emph{Thermus thermophilus} HB8 (taxids 146891 and 300852), obtained from the Codon Usage Database (Nakamura et al., 2000).}
  }
}

\references{
Dick, J. M., Yu, M., Tan, J. and Lu, A. (2019) Changes in carbon oxidation state of metagenomes along geochemical redox gradients. \emph{Front. Microbiol.} \bold{10}, 120. \doi{10.3389/fmicb.2019.00120}

Nakamura, Y., Gojobori, T. and Ikemura, T. (2000) Codon usage tabulated from international DNA sequence databases: status for the year 2000. \emph{Nucleic Acids Res.} \bold{28}, 292. \doi{10.1093/nar/28.1.292}

Rho, M., Tang, H. and Ye, Y. (2010) FragGeneScan: predicting genes in short and error-prone reads. \emph{Nucleic Acids Res.} \bold{38}, e191. \doi{10.1093/nar/gkq747}

Wood, D. E. and Salzberg, S. L. (2014) Kraken: ultrafast metagenomic sequence classification using exact alignments. \emph{Genome Biol.} \bold{15}, R46. \doi{10.1186/gb-2014-15-3-r46}
}

\examples{
gradox1()

mout <- gradoxS1()
pout <- gradoxS2()
gradox3(mout, pout)
}
