\encoding{UTF-8}
\name{JMDplots-package}
\alias{JMDplots-package}
\alias{JMDplots}
\docType{package}
\title{Plots from Papers by Jeffrey M. Dick}

\description{
This package contains data and code used to make the plots in different papers by the author.
The code is not identical to that used in the papers, but has been adapted from scripts to be usable in this package.
Each paper is identified with a project name.
These are summarized in the table below and are used for the installed directories under \code{extdata} and function names.
The specific function names have the figure numbers appended, as in e.g. \code{\link{gradox1}}.
}

\details{
  \itemize{
    \item \code{\link{gradox}} - redox gradients (Dick et al., 2019)
  }
  The data directories for each of these papers are described below.
}

\section{gradox}{
  The files were generated by processing source FASTA files using the "ARAST" workflow; see the paper and the Zenodo record (\doi{10.5281/zenodo.2314933}) for more information.
  In the first three directories, files whose names contain "MGD", "MGR", or "MGP" are derived from metegenomic dataset; those with "MTD", "MTR", or "MTP" are derived from metatranscriptomic datasets.

  \describe{
    \item{MGD}{DNA nucleobase compositions for groups of sequences sampled from source FASTA files that have been trimmed, normalized, and dereplicated. For metatranscriptomes, sequences classified as rRNA by FragGeneScan (Rho et al., 2010) were removed.}
    \item{MGR}{RNA nucleobase compositions for groups of sequences sampled from coding DNA FASTA files generated by FragGeneScan.}
    \item{MGP}{Protein amino acid compositions for groups of sequences sampled from protein FASTA files generated by FragGeneScan.}
    \item{one_percent}{Files generated by using Kraken (Wood and Salzberg, 2014) report files to calculate taxon abundances as a percentage of the number of classified reads, then extracting species that make up at least 1 percent of the classified reads.}
  }
}

\section{Requirements}{
Requires R 3.6.0 for the \code{gap.axis} argument in \code{\link{axis}}.
Requires CHNOSZ 1.3.3 for updated format of OBIGT data frame.
}

\references{
Dick, J. M., Yu, M., Tan, J. and Lu, A. (2019) Changes in carbon oxidation state of metagenomes along geochemical redox gradients. \emph{Front. Microbiol.} \bold{10}, 120. \doi{10.3389/fmicb.2019.00120}

Rho, M., Tang, H. and Ye, Y. (2010) FragGeneScan: predicting genes in short and error-prone reads. \emph{Nucleic Acids Res.} \bold{38}, e191. \doi{10.1093/nar/gkq747}

Wood, D. E. and Salzberg, S. L. (2014) Kraken: ultrafast metagenomic sequence classification using exact alignments. \emph{Genome Biol.} \bold{15}, R46. \doi{10.1186/gb-2014-15-3-r46}
}
