\encoding{UTF-8}
\name{JMDplots-package}
\alias{JMDplots-package}
\alias{JMDplots}
\docType{package}
\title{Plots from papers by Jeffrey M. Dick}

\description{
This package contains data and code used to make the plots in different papers by the author.
The code is not identical to that used in the papers, but has been adapted to be usable in this package.
Each paper is identified with a project name, as listed in the table below.
The plots are available in the corresponding vignettes in the package.
}

\details{
  \itemize{
    \item \code{\link{gradH2O}} - Hydration state of metagenomes in salinity gradients (Dick et al., 2019b)
    \item \code{\link{chnszten}} - CHNOSZ, ten years after first CRAN submission (Dick, 2019)
    \item \code{\link{gradox}} - Carbon oxidation state of metagenomes in redox gradients (Dick et al., 2019a)
    \item \code{aaaq} - Amino acid group additivity for ionized proteins (Dick et al., 2006)
  }
  These names are used for the vignettes and functions, as well as data directories under \code{extdata}.
  The function names have figure numbers appended, as in \code{\link{gradox1}}.
  Note that \code{aaaq} has a vignette but no functions.
  The data directories are described below.
}

\section{gradox and gradH2O}{
  The files were generated by processing source FASTA files using the "ARAST" workflow; see the paper (Dick et al., 2019a) and the Zenodo record (\doi{10.5281/zenodo.2314933}) for more information.
  In the first three directories, files whose names contain "MGD", "MGR", or "MGP" are derived from metegenomic datasets; those with "MTD", "MTR", or "MTP" are derived from metatranscriptomic datasets.

  \describe{
    \item{MGD}{DNA nucleobase compositions for groups of sequences sampled from source FASTA files that have been trimmed, normalized, and dereplicated. For metatranscriptomes, sequences classified as rRNA by FragGeneScan (Rho et al., 2010) were removed.}
    \item{MGR}{RNA nucleobase compositions for groups of sequences sampled from coding DNA FASTA files generated by FragGeneScan.}
    \item{MGP}{Protein amino acid compositions for groups of sequences sampled from protein FASTA files generated by FragGeneScan.}
    \item{one_percent}{Files generated by using Kraken (Wood and Salzberg, 2014) report files to calculate taxon abundances as a percentage of the number of classified reads, then extracting species that make up at least 1 percent of the classified reads.}
  }
}

\section{aaaq}{
  \describe{
    \item{BKM60_Fig7.csv}{Eh-pH values for normal, wet and waterlogged soils from Fig. 7 of Baas Becking et al. (1960).}
  }
}

\section{Other Data}{
Besides the above directories for the papers, there are additional directories with data to support examples or specific functions.

  \describe{
    \item{subsurface}{
      Metagenome-derived amino acid compositions for two subsurface environments.
      These files are provided to illustrate the usage of user-provided data sets; see \code{\link{subsurface}}.
    }

    \item{organisms}{
      Amino acid compositions of proteins from \emph{E. coli}, used in \code{\link{gradH2O1}}.
      See \code{\link{organisms}}.
    }

    \item{refseq}{
      Total amino acid compositions of all proteins for each taxon (bacteria, archaea, and viral) in the NCBI Reference Sequence database, based on RefSeq release 95 of 2019-07-12 (O'Leary et al., 2016).
      This directory contains the scripts used to process the RefSeq files and resulting CSV file with amino acid compositions.
    }
  }
}

\section{Requirements}{
Requires CHNOSZ 1.3.2 for functionality used in \code{\link{chnszten}}.
Requires canprot > 0.1.2 for updated plotting style in \code{\link[canprot]{diffplot}}.
}

\references{
Baas Becking, L. G. M., Kaplan, I. R. and Moore, D. (1960) Limits of the natural environment in terms of pH and oxidation-reduction potentials. \emph{Journal of Geology} \bold{68}, 243--284. \url{http://www.jstor.org/stable/30059218}

Dick, J. M., LaRowe, D. E. and Helgeson, H. C. (2006) Temperature, pressure, and electrochemical constraints on protein speciation: Group additivity calculation of the standard molal thermodynamic properties of ionized unfolded proteins. \emph{Biogeosciences} \bold{3}, 311--336. \doi{10.5194/bg-3-311-2006}

Dick, J. M. (2019) CHNOSZ: Thermodynamic calculations and diagrams for geochemistry. \emph{Front. Earth Sci.} \bold{7}:180. \doi{10.3389/feart.2019.00180}

Dick, J. M., Yu, M., Tan, J. and Lu, A. (2019a) Changes in carbon oxidation state of metagenomes along geochemical redox gradients. \emph{Front. Microbiol.} \bold{10}, 120. \doi{10.3389/fmicb.2019.00120}

Dick, J. M., Tan, J. and Yu, M. (2019b) Geochemistry of metagenomes: Distinct compositional signatures of environmental redox and salinity gradients. \emph{(preprint)}

O'Leary, Nuala A. et al. (2016) Reference sequence (RefSeq) database at NCBI: current status, taxonomic expansion, and functional annotation. \emph{Nucleic Acids Res.} \bold{44}, D733-D745. \doi{10.1093/nar/gkv1189}

Rho, M., Tang, H. and Ye, Y. (2010) FragGeneScan: predicting genes in short and error-prone reads. \emph{Nucleic Acids Res.} \bold{38}, e191. \doi{10.1093/nar/gkq747}

Wood, D. E. and Salzberg, S. L. (2014) Kraken: ultrafast metagenomic sequence classification using exact alignments. \emph{Genome Biol.} \bold{15}, R46. \doi{10.1186/gb-2014-15-3-r46}
}
