\encoding{UTF-8}
\name{JMDplots-package}
\alias{JMDplots-package}
\alias{JMDplots}
\docType{package}
\title{Plots from papers by Jeffrey M. Dick}

\description{
This package contains data and code used to make the plots in different papers by the author.
Each paper is identified with a project name, as listed in the table below.
The plots are available in the corresponding vignettes in the package.

Note that the code in this package is not identical to that used in the papers, but has been minimally modified to be included in this package.
Please refer to the supporting information of the papers for the original scripts.
}

\details{
  \itemize{
    \item \code{\link{canH2O}} - Water as a reactant in the differential expression of proteins in cancer (Dick, 2020)
    \item \code{\link{gradH2O}} - Hydration state of metagenomes in salinity gradients (Dick et al., 2020)
    \item \code{\link{chnszten}} - CHNOSZ, ten years after first CRAN submission (Dick, 2019)
    \item \code{\link{gradox}} - Carbon oxidation state of metagenomes in redox gradients (Dick et al., 2019)
    \item \code{\link{aoscp}} - Average oxidation state of carbon in proteins (Dick, 2014)
    \item \code{\link{scsc}} - Subcellular locations of \emph{Saccharomyces cerevisiae} (Dick, 2009)
    \item \code{\link{aaaq}} - Amino acid group additivity for ionized proteins (Dick et al., 2006)
  }
  These names are used for the vignettes and functions; the function names have figure numbers appended, as in \code{\link{gradox1}}.
  Data for each of the papers are stored in the corresponding directories under \code{extdata}; see the documentation page for each paper for more details.
  There are some other directories, described below.

  \describe{
    \item{\code{subsurface}}{
      Metagenome-derived amino acid compositions for two subsurface environments.
      These files are provided to illustrate the usage of user-provided data sets; see \code{\link{subsurface}}.
    }
    \item{\code{organisms}}{
      Amino acid compositions of yeast and \emph{E. coli}, and subcellular location data for yeast.
      See \code{\link{yeast}}, \code{\link{scsc}} and \code{\link{gradH2O}}.
    }

    \item{\code{refseq}}{
      Total amino acid compositions of all proteins for each taxon (bacteria, archaea, and viral) in the NCBI Reference Sequence database, based on RefSeq release 95 of 2019-07-12 (O'Leary et al., 2016).
      This directory contains the scripts used to process the RefSeq files and resulting CSV file with amino acid compositions.
      \describe{
        \item{\code{protein_refseq.csv.xz}}{
          The columns are arranged as follows: \code{protein} (\dQuote{refseq}), \code{organism} (taxonomic id), \code{ref} (organism name), \code{abbrv} (empty), \code{chains} (number of protein sequences for this organism).
          The remaining columns have the total counts of amino acids.
          Used in \code{\link{gradH2O}}.
        }
      }
    }

    \item{\code{vignette_output}}{
      This directory has CSV files generated by the vignettes TCGA.Rmd and HPA.Rmd.
      These files are used in \code{\link{canH2O2}}.
    }
  }

}

\section{Other Papers}{

This package does not include plots from \emph{all} papers by the author.
Many plots from the papers by Dick and Shock (2011 and 2013) are available in the vignette \dQuote{Hot-spring proteins in CHNOSZ} (\url{../../CHNOSZ/doc/hotspring.pdf}).
See also \code{demo("bison", "CHNOSZ")} for a colorful version of Figure 1 of Dick and Shock (2013) (\ZC plot for phyla in Bison Pool).
The plots from Dick (2017) are in the vignettes for \pkg{canprot} (\url{../../canprot/doc/index.html}).
See \code{demo("bugstab", "CHNOSZ")} for Figure 4 of Dick (2016) (relative stability diagram of microbial proteins in colorectal cancer) and \code{oldvignettes/colorectal.Rmd} in \pkg{canprot} for Figure 6 of Dick (2016) (relative stability diagram for differentially expressed proteins in various colorectal cancer datasets).

}

\section{Acknowledgements}{
The results shown here are in part based upon data generated by the TCGA Research Network (\url{https://www.cancer.gov/tcga}) and Human Protein Atlas (\url{https://www.proteinatlas.org/}).
}

\references{
Dick, J. M., LaRowe, D. E. and Helgeson, H. C. (2006) Temperature, pressure, and electrochemical constraints on protein speciation: Group additivity calculation of the standard molal thermodynamic properties of ionized unfolded proteins. \emph{Biogeosciences} \bold{3}, 311--336. \doi{10.5194/bg-3-311-2006}

Dick, J. M. (2009) Calculation of the relative metastabilities of proteins in subcellular compartments of \emph{Saccharomyces cerevisiae}. \emph{BMC Syst. Biol.} \bold{3}, 75. \doi{10.1186/1752-0509-3-75}

Dick, J. M. and Shock, E. L. (2011) Calculation of the relative chemical stabilities of proteins as a function of temperature and redox chemistry in a hot spring. \emph{PLoS ONE} \bold{6}, e22782. \doi{10.1371/journal.pone.0022782}

Dick, J. M. and Shock, E. L. (2013) A metastable equilibrium model for the relative abundance of microbial phyla in a hot spring. \emph{PLoS ONE} \bold{8}, e72395. \doi{10.1371/journal.pone.0072395}

Dick, J. M. (2014) Average oxidation state of carbon in proteins. \emph{J. R. Soc. Interface} \bold{11}, 20131095. \doi{10.1098/rsif.2013.1095}

Dick, J. M. (2016) Proteomic indicators of oxidation and hydration state in colorectal cancer. \emph{PeerJ} \bold{4}, e2238. \doi{10.7717/peerj.2238}

Dick, J. M. (2017) Chemical composition and the potential for proteomic transformation in cancer, hypoxia, and hyperosmotic stress. \emph{PeerJ} \bold{5}, e3421 \doi{10.7717/peerj.3421}

Dick, J. M., Yu, M., Tan, J. and Lu, A. (2019) Changes in carbon oxidation state of metagenomes along geochemical redox gradients. \emph{Front. Microbiol.} \bold{10}, 120. \doi{10.3389/fmicb.2019.00120}

Dick, J. M. (2019) CHNOSZ: Thermodynamic calculations and diagrams for geochemistry. \emph{Front. Earth Sci.} \bold{7}:180. \doi{10.3389/feart.2019.00180}

Dick, J. M., Tan, J. and Yu, M. (2020) Geochemistry of metagenomes: Redox and salinity gradients and the chemical compositions of proteins. \emph{(preprint)}

Dick, J. M. (2020) Water as a reactant in the differential expression of proteins in cancer. \emph{(preprint)}

O'Leary, N. A. et al. (2016) Reference sequence (RefSeq) database at NCBI: current status, taxonomic expansion, and functional annotation. \emph{Nucleic Acids Res.} \bold{44}, D733-D745. \doi{10.1093/nar/gkv1189}
}
